+++ packages/api-client/src/index.ts
@@ -1,11 +1,5 @@
 /* eslint-disable @typescript-eslint/no-namespace */
 
-export namespace SECFund {
-  export type Paths = import("./sec-fund").paths;
-  export type Components = import("./sec-fund").components;
-  export type Operations = import("./sec-fund").operations;
-}
-
 export namespace SECV2 {
   export type Paths = import("./sec-v2").paths;
   export type Components = import("./sec-v2").components;